requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
psycopg2>=2.9.9
pandas>=2.0.0
openpyxl>=3.1.0
pydantic>=2.7.0
python-dotenv>=1.0.0
gunicorn